# openapi.yaml
...
openapi: 3.1.0
info:
  title: photolib
  description: "\nA local, private photo library. Natural-language search over your photos,\nface search\
    \ and recognition, and date/location browsing \u2014 all running on your\nown machine with no network\
    \ calls.\n"
  version: 2.0.0
paths:
  /api/v1/search:
    post:
      tags:
      - search
      summary: Search
      description: 'Combined semantic + person + date + location search.


        With no query this is a plain browse, which is why an empty search is

        cheap: it never touches the embedding model or the ANN index.'
      operationId: search_api_v1_search_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - search
      summary: Search Get
      description: GET form of /search, so a search is a shareable, bookmarkable URL.
      operationId: search_get_api_v1_search_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Natural-language query
          title: Q
        description: Natural-language query
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 1000
            minimum: 1
          - type: 'null'
          title: Per Page
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: relevance
          title: Sort
      - name: people
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Comma-separated person ids
          title: People
        description: Comma-separated person ids
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/faces/unassigned:
    get:
      tags:
      - faces
      summary: Unassigned
      description: "Faces with no identity yet, best-quality first \u2014 the review queue."
      operationId: unassigned_api_v1_faces_unassigned_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 120
          title: Limit
      - name: min_quality
        in: query
        required: false
        schema:
          type: number
          maximum: 1.0
          minimum: 0.0
          default: 0.3
          title: Min Quality
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/search:
    post:
      tags:
      - faces
      summary: Search Faces
      description: 'Find faces that look like a given face or person.


        This is the "search by face" path: it queries the face vector index

        directly, so it finds people the clusterer never linked, including in

        photos where they were never tagged.'
      operationId: search_faces_api_v1_faces_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceSearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/search/by-upload:
    post:
      tags:
      - faces
      summary: Search Faces By Upload
      description: "\"Who is this?\" \u2014 upload a photo and match its largest face."
      operationId: search_faces_by_upload_api_v1_faces_search_by_upload_post
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 300
          minimum: 1
          default: 60
          title: Limit
      - name: min_similarity
        in: query
        required: false
        schema:
          type: number
          maximum: 1.0
          minimum: -1.0
          default: 0.3
          title: Min Similarity
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_search_faces_by_upload_api_v1_faces_search_by_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/assign:
    post:
      tags:
      - faces
      summary: Assign Faces
      description: 'Attach faces to a person, or create a new person from them.


        Assignments made here are marked confirmed, so a later recluster treats

        them as ground truth rather than undoing the correction.'
      operationId: assign_faces_api_v1_faces_assign_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignFacesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/detach:
    post:
      tags:
      - faces
      summary: Detach Faces
      description: "\"That isn't them\" \u2014 return faces to the unassigned pool."
      operationId: detach_faces_api_v1_faces_detach_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetachFacesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/{face_id}:
    get:
      tags:
      - faces
      summary: Get Face
      operationId: get_face_api_v1_faces__face_id__get
      parameters:
      - name: face_id
        in: path
        required: true
        schema:
          type: integer
          title: Face Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/faces/{face_id}/crop:
    get:
      tags:
      - faces
      summary: Get Face Crop
      description: The cropped face, cut from the original photo and cached on first view.
      operationId: get_face_crop_api_v1_faces__face_id__crop_get
      parameters:
      - name: face_id
        in: path
        required: true
        schema:
          type: integer
          title: Face Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssignFacesRequest:
      properties:
        face_ids:
          items:
            type: integer
          type: array
          minItems: 1
          title: Face Ids
        person_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Person Id
          description: Omit to create a new person from these faces
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Name for a newly created person
      type: object
      required:
      - face_ids
      title: AssignFacesRequest
...
    Body_search_faces_by_upload_api_v1_faces_search_by_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_search_faces_by_upload_api_v1_faces_search_by_upload_post
    DetachFacesRequest:
      properties:
        face_ids:
          items:
            type: integer
          type: array
          minItems: 1
          title: Face Ids
      type: object
      required:
      - face_ids
      title: DetachFacesRequest
...
    FaceSearchRequest:
      properties:
        face_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Face Id
        person_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Person Id
        limit:
          type: integer
          maximum: 500.0
          minimum: 1.0
          title: Limit
          default: 60
        min_similarity:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Min Similarity
          default: 0.3
      type: object
      title: FaceSearchRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ImageSummary:
      properties:
        favorite:
          type: boolean
          title: Favorite
          default: false
        rating:
          type: integer
          title: Rating
          default: 0
        quality_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Quality Score
        quality_reasons:
          items:
            type: string
          type: array
          title: Quality Reasons
        image_id:
          type: integer
          title: Image Id
        filename:
          type: string
          title: Filename
          default: ''
        taken_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Taken At
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lon:
          anyOf:
          - type: number
          - type: 'null'
          title: Lon
        place:
          type: string
          title: Place
          default: ''
        people_ids:
          items:
            type: integer
          type: array
          title: People Ids
        face_count:
          type: integer
          title: Face Count
          default: 0
        width:
          type: integer
          title: Width
          default: 0
        height:
          type: integer
          title: Height
          default: 0
        media_type:
          type: string
          title: Media Type
          default: image
        duration_ms:
          type: integer
          title: Duration Ms
          default: 0
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
        text_match:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Text Match
      type: object
      required:
      - image_id
      title: ImageSummary
...
    SearchRequest:
      properties:
        favorites_only:
          type: boolean
          title: Favorites Only
          default: false
        min_rating:
          type: integer
          maximum: 5.0
          minimum: 0.0
          title: Min Rating
          default: 0
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
          description: Natural-language description
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Date
        people_ids:
          items:
            type: integer
          type: array
          title: People Ids
        people_mode:
          type: string
          enum:
          - any
          - all
          title: People Mode
          default: any
        has_location:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Location
        has_faces:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Faces
        folder:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder
        camera:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera
        untagged_only:
          type: boolean
          title: Untagged Only
          default: false
        near_lat:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Near Lat
        near_lon:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Near Lon
        near_km:
          type: number
          maximum: 20000.0
          exclusiveMinimum: 0.0
          title: Near Km
          default: 1.0
        media:
          anyOf:
          - type: string
            enum:
            - image
            - video
          - type: 'null'
          title: Media
        sort:
          type: string
          enum:
          - relevance
          - quality
          - date_desc
          - date_asc
          - added_desc
          - random
          title: Sort
          default: relevance
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        per_page:
          anyOf:
          - type: integer
            maximum: 1000.0
            minimum: 1.0
          - type: 'null'
          title: Per Page
        min_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: -1.0
          - type: 'null'
          title: Min Score
          description: Drop semantic matches below this cosine similarity
      type: object
      title: SearchRequest
    SearchResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
        took_ms:
          type: number
          title: Took Ms
          default: 0.0
        scored:
          type: boolean
          title: Scored
          default: false
        results:
          items:
            $ref: '#/components/schemas/ImageSummary'
          type: array
          title: Results
      type: object
      required:
      - total
      - page
      - per_page
      - results
      title: SearchResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError